feat(messages): increase file upload limit to 2 GB

+++ services/api/internal/messages/handler_response.go
@@ -40,7 +40,7 @@ type fileResponse struct {
 	URL         string `json:"url"`
 	Filename    string `json:"filename"`
 	ContentType string `json:"content_type"`
-	ByteSize    uint32 `json:"byte_size"`
+	ByteSize    uint64 `json:"byte_size"`
 }
 
 func responseFromMessage(message Message) messageResponse {
